+++ sdk/mx.sdk/mx_sdk_vm_impl.py
@@ -3555,6 +3555,31 @@ def mx_register_dynamic_suite_constituents(register_project, register_distributi
     :type register_project: (mx.Project) -> None
     :type register_distribution: (mx.Distribution) -> None
     """
+    def _release_version():
+        version = _suite.release_version()
+        if version.endswith('-dev'):
+            version = version[:-len('-dev')]
+        return version
+
+    string_substitutions = mx_subst.SubstitutionEngine(mx_subst.SubstitutionEngine(mx_subst.path_substitutions))
+    string_substitutions.register_no_arg('version', _release_version)
+    attrs = {
+        'description': 'SDK version file.',
+        'maven': False,
+    }
+    register_distribution(mx.LayoutDirDistribution(
+        suite=_suite,
+        name='VERSION',
+        deps=[],
+        layout={
+            'version': 'string:<version>'
+        },
+        path=None,
+        platformDependent=False,
+        theLicense=None,
+        string_substitutions=string_substitutions,
+        **attrs
+    ))
     main_dists = {
         'graalvm': [],
         'graalvm_installables': [],