+++ sdk/src/org.graalvm.polyglot/src/org/graalvm/polyglot/impl/AbstractPolyglotImpl.java
@@ -586,6 +586,10 @@ public abstract class AbstractPolyglotImpl {
         return getNext().copyResources(targetFolder, components);
     }
 
+    public String getTruffleVersion() {
+        return null;
+    }
+
     /**
      * Marker base class for native-image.
      */