+++ truffle/mx.truffle/suite.py
@@ -1364,8 +1364,7 @@ suite = {
       "dependencies" : [
         "com.oracle.truffle.compiler",
       ],
-      "distDependencies" : [
-      ],
+      "distDependencies" : [],
       "description" : "Truffle compiler API.",
       "maven": {
           "tag": ["default", "public"],
@@ -1424,6 +1423,16 @@ suite = {
           "artifactId": "truffle-runtime",
           "tag": ["default", "public"],
       },
+    },
+
+    "TRUFFLE_API_VERSION": {
+      "type": "dir",
+      "platformDependent": False,
+      "layout": {
+        "META-INF/graalvm/org.graalvm.truffle/version": "dependency:sdk:VERSION/version",
+      },
+      "description": "Truffle API version.",
+      "maven": False,
     },
 
     "TRUFFLE_API" : {
@@ -1511,9 +1520,10 @@ suite = {
         "com.oracle.truffle.polyglot",
         "com.oracle.truffle.host",
         "com.oracle.truffle.api.staticobject",
+        "TRUFFLE_API_VERSION",
       ],
       "distDependencies" : [
-        "sdk:POLYGLOT"
+        "sdk:POLYGLOT",
       ],
       "description" : "Truffle is a multi-language framework for executing dynamic languages\nthat achieves high performance when combined with Graal.",
       "javadocType": "api",