+++ truffle/src/com.oracle.truffle.compiler/src/com/oracle/truffle/compiler/hotspot/libgraal/TruffleToLibGraal.java
@@ -64,6 +64,7 @@ public @interface TruffleToLibGraal {
     enum Id {
         DoCompile,
         GetCompilerConfigurationFactoryName,
+        GetCompilerVersion,
         GetDataPatchesCount,
         GetExceptionHandlersCount,
         GetInfopoints,